+++ app/src/admin/installer.controller.js
@@ -3,10 +3,13 @@ angular.module('AmpersandApp')
     $scope.installing = false;
     $scope.installed = false;
     
-    $scope.install = function(defPop){
+    $scope.install = function(defPop, ignoreInvariantRules){
         $scope.installing = true;
         $scope.installed = false;
-        Restangular.one('admin/installer').get({defaultPop : defPop}).then(function(data) {
+        Restangular
+        .one('admin/installer')
+        .get({defaultPop : defPop, ignoreInvariantRules : ignoreInvariantRules})
+        .then(function(data) {
             data = data.plain();
             NotificationService.updateNotifications(data);
             NavigationBarService.refreshNavBar();